+++ docs/changelog.md
@@ -1,7 +1,7 @@
 ﻿# Kapoin changelog
 
 ### Version 0.0.2.1
-* (work in progress)
+* Add example contracts 'North of Sixty' (C60) and 'Take a swim' (C70).
 * Add main loop for progress tracking, and add node data to persistence file.
 * Add game setting 'Track Progress'.